src/requirements.txt
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.7.1
cairocffi==1.0.2
CairoSVG==2.3.1
certifi==2019.3.9
cffi==1.15.1
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
django>=2.2.11
django-anysign==1.4
django-crispy-forms==1.8.1
django-filter==21.1
django-heroku==0.3.1
django-taggit==3.1.0
dj-static==0.0.6
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
nltk==3.8.1
pycodestyle==2.5.0
pycparser==2.19
pydocusign==2.2
pylint==2.3.1
Pyphen==0.9.5
python-dateutil==2.8.0
python-decouple==3.3
py-trello==0.18.0
pytz==2019.1
requests==2.21.0
scikit-learn==1.2.1
six==1.16.0
soupsieve==1.9.1
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.5.4
urllib3==1.24.3
WeasyPrint==47
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.11.1
